backend/requirements.txt
uvicorn
fastapi
sqlalchemy
python-dotenv
uuid
pydantic